Updated Menu & Program - Started Client
Started writing the new client class to replace the standard TCP in the main class of Program.cs. Updated the Byte array size in Program.cs & changed the return value of one of the functions in HanxMenu. More updates & documentation to come. None of this had been debuged since it was written on my compuer w/o a compiler.